-Completadas funcionalidades de la gestión de Aplicaciones, Usuarios y Certificados. -Correcciones relacionadas con la persistencia en la Gestión de Logs: Posibilidad de usar Oracle o MySql dependiendo del dialecto detectado. -Tareas relacionadas con la documentación javadoc y eliminar recursos innecesarios (en curso).

+++ fire-admin-web/src/main/java/es/gob/fire/web/rest/controller/CertificateRestController.java
@@ -1,3 +1,27 @@
+/* 
+/*******************************************************************************
+ * Copyright (C) 2018 MINHAFP, Gobierno de España
+ * This program is licensed and may be used, modified and redistributed under the  terms
+ * of the European Public License (EUPL), either version 1.1 or (at your option)
+ * any later version as soon as they are approved by the European Commission.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and
+ * more details.
+ * You should have received a copy of the EUPL1.1 license
+ * along with this program; if not, you may find it at
+ * http:joinup.ec.europa.eu/software/page/eupl/licence-eupl
+ ******************************************************************************/
+
+/** 
+ * <b>File:</b><p>es.gob.fire.web.rest.controller.CertificateRestController.java.</p>
+ * <b>Description:</b><p>Class that manages the REST requests related to the Certificate administration and JSON communication.</p>
+  * <b>Project:</b><p>Application for signing documents of @firma suite systems</p>
+ * <b>Date:</b><p>22/01/2021.</p>
+ * @author Gobierno de España.
+ * @version 1.0, 22/01/2021.
+ */
 package es.gob.fire.web.rest.controller;
 
 import java.io.IOException;
@@ -37,12 +61,13 @@ import es.gob.fire.i18n.IWebViewMessages;
 import es.gob.fire.i18n.Language;
 import es.gob.fire.persistence.dto.ApplicationCertDTO;
 import es.gob.fire.persistence.dto.CertificateDTO;
+import es.gob.fire.persistence.entity.Application;
 import es.gob.fire.persistence.entity.Certificate;
 import es.gob.fire.persistence.service.IApplicationService;
 import es.gob.fire.persistence.service.ICertificateService;
 
 /**
- * <p>Class that manages the REST requests related to the Users administration and JSON communication.</p>
+ * <p>Class that manages the REST requests related to the Certificate administration and JSON communication.</p>
  * <b>Project:</b><p>Application for signing documents of @firma suite systems.</p>
  * @version 1.0, 24/12/2020.
  */
@@ -155,15 +180,22 @@ public class CertificateRestController {
 	 */
 	@JsonView(DataTablesOutput.View.class)
 	@RequestMapping(path = "/deletecertificate", method = RequestMethod.POST)
-	public String deleteCertificate(@RequestParam(FIELD_ID_CERTIFICATE) final Long idCertificate, @RequestParam(FIELD_ROW_INDEX_CERTIFICATE) final String index) {
+	public String deleteCertificate(@RequestParam(FIELD_ID_CERTIFICATE) final Long idCertificate,
+			@RequestParam(FIELD_ROW_INDEX_CERTIFICATE) final String index) {
 		String result = index;
-		
-		try {
-			certificateService.deleteCertificate(idCertificate);
-		} catch (Exception e) {
-			result = "-1";
+
+		List<Application> aplicacionesCert = appService.getByIdCertificado(idCertificate);
+
+		if (aplicacionesCert != null || aplicacionesCert.size() > 0) {
+			result = "error.No se ha podido borrar el certificado, tiene aplicaciones asociadas.";
+		} else {
+
+			try {
+				certificateService.deleteCertificate(idCertificate);
+			} catch (Exception e) {
+				result = "-1";
+			}
 		}
-		
 		return result;
 	}
